backend/requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
fastapi==0.111.1
uvicorn[standard]==0.30.1
websockets==12.0